os/bluestore: Moved NCB related code to FileFreelistManager
For now, moved code is still a part of BlueStore class.
In future it will became part of FileFreelistManager.
